requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8
Pillow==11.1.0
django-filter==24.3
